+++ bot.py
@@ -1,85 +1,108 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """Methods for the bot functionality."""
-import sys
+import html
+import json
+import logging
 import traceback
-from typing import Optional
-from telegram import Update, InlineKeyboardMarkup, InlineKeyboardButton
-from telegram.ext import (CallbackContext, Dispatcher, MessageHandler, PrefixHandler,
-                          CommandHandler, CallbackQueryHandler, Filters)
-from telegram.utils.helpers import mention_html
-from emoji import emojize
+from typing import List, Optional, cast
 
-HOMEPAGE: str = 'https://hirschheissich.gitlab.io/memoriae-bot/'
+from telegram import Chat, InlineKeyboardButton, InlineKeyboardMarkup, Message, Update
+from telegram.error import BadRequest, Forbidden
+from telegram.ext import (
+    Application,
+    CallbackQueryHandler,
+    CommandHandler,
+    ContextTypes,
+    MessageHandler,
+    PrefixHandler,
+    filters,
+)
+
+HOMEPAGE: str = "https://Bibo-Joshi.github.io/memoriae-bot/"
 """:obj:`str`: Homepage of this bot."""
 ADMIN: int = -1
 """:obj:`int`: Chat ID of the admin. Needs to be overridden in main.py"""
-CALLBACK_NO: str = 'dont_postpone'
+CALLBACK_NO: str = "dont_postpone"
 """:obj:`str`: Callback data for not postponing reminders."""
 
 
-def info(update: Update, context: CallbackContext) -> None:
+async def info(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
     """
     Returns some info about the bot.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        _: The callback context as provided by the application.
     """
-    text = emojize(
-        ('I\'m the <b>Memoriae Bot</b> and I\'m a thin wrapper for Telegrams timed messages. My '
-         'profession is postponing reminders. Just send me a timed message and follow along.'
-         '\n\nTo learn more about me, please visit my homepage '
-         ':slightly_smiling_face:.'),
-        use_aliases=True)
+    text = (
+        "I'm the <b>Memoriae Bot</b> and I'm a thin wrapper for Telegrams timed messages. My "
+        "profession is postponing reminders. Just send me a timed message and follow along."
+        "\n\nTo learn more about me, please visit my homepage 🙂."
+    )
 
-    keyboard = InlineKeyboardMarkup.from_button(
-        InlineKeyboardButton(emojize('Memoriae Bot :robot_face:', use_aliases=True), url=HOMEPAGE))
+    reply_markup = InlineKeyboardMarkup.from_button(
+        InlineKeyboardButton("Memoriae Bot 🤖", url=HOMEPAGE)
+    )
 
-    update.message.reply_text(text, reply_markup=keyboard)
+    await cast(Message, update.message).reply_text(text, reply_markup=reply_markup)
 
 
-def error(update: Update, context: CallbackContext):
-    """
-    Informs the originator of the update that an error occured and forwards the traceback to the
-    admin.
+async def error(update: object, context: ContextTypes.DEFAULT_TYPE) -> None:
+    """Informs the originator of the update that an error occurred and forwards the traceback to
+    the admin.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
     """
+    admin_id = ADMIN
+    logger = logging.getLogger(__name__)
+
+    # Log the error before we do anything else, so we can see it even if something breaks.
+    logger.error(msg="Exception while handling an update:", exc_info=context.error)
+
+    if (
+        isinstance(context.error, Forbidden)
+        or (isinstance(context.error, BadRequest) and "Query is too old" in str(context.error))
+        or context.error is None
+    ):
+        return
+
     # Inform sender of update, that something went wrong
-    if update.effective_message:
-        text = emojize('Something went wrong :worried:. I informed the admin :nerd_face:.',
-                       use_aliases=True)
-        update.effective_message.reply_text(text)
+    if isinstance(update, Update) and update.effective_message:
+        text = "Something went wrong 😟. I informed the admin 🤓."
+        await update.effective_message.reply_text(text)
 
     # Get traceback
-    trace = ''.join(traceback.format_tb(sys.exc_info()[2]))
+    tb_list = traceback.format_exception(None, context.error, context.error.__traceback__)
+    tb_string = "".join(tb_list)
 
-    # Gather information from the update
-    payload = ''
-    if update.effective_user:
-        payload += ' with the user {}'.format(
-            mention_html(update.effective_user.id, update.effective_user.first_name))
-    if update.effective_chat:
-        payload += f' within the chat <i>{update.effective_chat.title}</i>'
-        if update.effective_chat.username:
-            payload += f' (@{update.effective_chat.username})'
-    if update.poll:
-        payload += f' with the poll id {update.poll.id}.'
-    text = f'Hey.\nThe error <code>{context.error}</code> happened{payload}. The full ' \
-           f'traceback:\n\n<code>{trace}</code>'
+    # Build the message with some markup and additional information about what happened.
+    update_str = update.to_dict() if isinstance(update, Update) else str(update)
+    message_1 = (
+        f"An exception was raised while handling an update\n\n"
+        f"<pre>update = {html.escape(json.dumps(update_str, indent=2, ensure_ascii=False))}</pre>"
+    )
+    message_2 = f"<pre>{html.escape(tb_string)}</pre>"
 
-    # Send to admin
-    context.bot.send_message(ADMIN, text)
+    # Finally, send the messages
+    # We send update and traceback in two parts to reduce the chance of hitting max length
+    sent_message = await context.bot.send_message(chat_id=admin_id, text=message_1)
+    try:
+        await sent_message.reply_html(message_2)
+    except BadRequest as exc:
+        if "too long" not in str(exc):
+            raise exc
+        message = (
+            f"Hey.\nThe error <code>{html.escape(str(context.error))}</code> happened."
+            f" The traceback is too long to send, but it was written to the log."
+        )
+        await context.bot.send_message(chat_id=admin_id, text=message)
 
-    # we raise the error again, so the logger module catches it.
-    raise  # pylint: disable=misplaced-bare-raise
 
-
-def keyboard(update: Optional[Update] = None,
-             message_id: Optional[str] = None) -> InlineKeyboardButton:
+def keyboard(
+    update: Optional[Update] = None, message_id: Optional[str] = None
+) -> InlineKeyboardMarkup:
     """
     Builds the keyboard to append to postpone messages. Exactly one of the optionals mey be passed.
 
@@ -90,87 +113,98 @@ def keyboard(update: Optional[Update] = None,
             postponed reminder.
     """
     if update is not None and message_id is not None:
-        raise ValueError('Only one optional argument may be passed.')
+        raise ValueError("Only one optional argument may be passed.")
 
-    if update is not None:
+    id_: Optional[str] = None
+    if update is not None and update.effective_message is not None:
         id_ = str(update.effective_message.message_id)
     if message_id is not None:
         id_ = message_id
 
-    return InlineKeyboardMarkup.from_row([
-        InlineKeyboardButton(text=emojize('No :heavy_check_mark:', use_aliases=True),
-                             callback_data=CALLBACK_NO),
-        InlineKeyboardButton(text=emojize('Yes :clock7:', use_aliases=True),
-                             switch_inline_query_current_chat=id_),
-    ])
+    return InlineKeyboardMarkup.from_row(
+        [
+            InlineKeyboardButton(text="No ✔️", callback_data=CALLBACK_NO),
+            InlineKeyboardButton(
+                text="Yes 🕖",
+                switch_inline_query_current_chat=id_,
+            ),
+        ]
+    )
 
 
-def answer_reminder(update: Update, context: CallbackContext):
+async def answer_reminder(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
     """
-    Answers a reminder and asks, whether to postpone it.
+    Answers a reminder and asks whether to postpone it.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        _: The callback context as provided by the application.
     """
-    update.message.reply_text('You need that again?', reply_markup=keyboard(update=update))
+    await cast(Message, update.message).reply_text(
+        "You need that again?", reply_markup=keyboard(update=update)
+    )
 
 
-def answer_postponed_reminder(update: Update, context: CallbackContext):
+async def answer_postponed_reminder(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
     """
-    Deletes the auxiliary reminder and posts the original one. Also asks, whether to postpone it
+    Deletes the auxiliary reminder and posts the original one. Also asks whether to postpone it
     again.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the application.
     """
-    update.message.delete()
+    await cast(Message, update.message).delete()
 
-    message_id = context.args[0]
-    reminder = context.bot.forward_message(chat_id=update.effective_chat.id,
-                                           from_chat_id=update.effective_chat.id,
-                                           message_id=message_id,
-                                           disable_notification=True)
+    message_id = cast(List[str], context.args)[0]
+    chat = cast(Chat, update.effective_chat)
+    reminder = await context.bot.forward_message(
+        chat_id=chat.id,
+        from_chat_id=chat.id,
+        message_id=int(message_id),
+        disable_notification=True,
+    )
 
-    reminder.reply_text('You need that again?',
-                        reply_markup=keyboard(message_id=message_id),
-                        disable_notification=True)
+    await reminder.reply_text(
+        "You need that again?",
+        reply_markup=keyboard(message_id=message_id),
+        disable_notification=True,
+    )
 
 
-def delete_message(update: Update, context: CallbackContext):
+async def delete_message(update: Update, _: ContextTypes.DEFAULT_TYPE) -> None:
     """
     Deletes the message of the incoming callback query.
 
     Args:
         update: The Telegram update.
-        context: The callback context as provided by the dispatcher.
+        _: The callback context as provided by the application.
     """
-    update.callback_query.message.delete()
+    await update.callback_query.message.delete()  # type: ignore[union-attr]
 
 
-def register_dispatcher(disptacher: Dispatcher) -> None:
+async def register_application(application: Application) -> None:
     """
     Adds handlers and sets up jobs. Convenience method to avoid doing that all in the main script.
 
     Args:
-        disptacher: The :class:`telegram.ext.Dispatcher`.
+        application: The :class:`telegram.ext.Dispatcher`.
     """
-    dp = disptacher
+    app = application
 
     # error handler
-    dp.add_error_handler(error)
+    app.add_error_handler(error)
 
     # basic command handlers
-    dp.add_handler(CommandHandler(['start', 'help'], info))
+    app.add_handler(CommandHandler(["start", "help"], info))
 
     # Handle postponed reminders
-    dp.add_handler(
-        PrefixHandler(prefix='@', command=dp.bot.username, callback=answer_postponed_reminder))
+    app.add_handler(
+        PrefixHandler(prefix="@", command=app.bot.username, callback=answer_postponed_reminder)
+    )
 
     # Handle first time reminders
-    dp.add_handler(MessageHandler(Filters.all, answer_reminder))
+    app.add_handler(MessageHandler(filters.ALL, answer_reminder))
 
-    # Delete postponing inqueries, when not needed
-    dp.add_handler(
-        CallbackQueryHandler(callback=delete_message, pattern='^{}$'.format(CALLBACK_NO)))
+    # Delete postponing queries, when not needed
+    app.add_handler(CallbackQueryHandler(callback=delete_message, pattern=f"^{CALLBACK_NO}$"))