+++ docs/source/conf.py
@@ -12,14 +12,15 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'Memoriae Bot'
-copyright = '2020, Hinrich Mahler'
-author = 'Hinrich Mahler'
+project = "Memoriae Bot"
+copyright = "2020-2023, Hinrich Mahler"
+author = "Hinrich Mahler"
 
 
 # -- General configuration ---------------------------------------------------
@@ -28,29 +29,30 @@ author = 'Hinrich Mahler'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.coverage',
-    'sphinx_autodoc_typehints',
-    'sphinx_rtd_theme'
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.coverage",
+    "sphinx_autodoc_typehints",
 ]
 
 # Use intersphinx to reference the python-telegram-bot docs
-intersphinx_mapping = {'telegram': ('https://python-telegram-bot.readthedocs.io/en/latest/', None),
-                       'https://docs.python.org/': None}
+intersphinx_mapping = {
+    "telegram": ("https://python-telegram-bot.readthedocs.io/en/stable/", None),
+    "https://docs.python.org/": None,
+}
 
 # Include special members in doc
 napoleon_include_special_with_doc = True
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -62,33 +64,21 @@ exclude_patterns = []
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'sphinx_rtd_theme'
+html_theme = "furo"
 html_theme_options = {
-    'style_nav_header_background': '#222426',
+    "navigation_with_keys": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = '../../logo/memoriae-bot-no-background.png'
+html_logo = "../../logo/memoriae-bot-no-background.png"
 
 # The name of an image file (relative to this directory) to use as a favicon of
 # the docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
-html_favicon = '../../logo/memoriae-bot-icon.ico'
-
-
-# Link to the correct file
-html_context = {
-    "display_gitlab": True,
-    "gitlab_host": "gitlab.com",
-    "gitlab_user": "HirschHeissIch",
-    "gitlab_repo": 'memoriae-bot',
-    "gitlab_version": "master",
-    "conf_py_path": "/docs/source/",
-    "source_suffix": '.rst',
-}
+html_favicon = "../../logo/memoriae-bot-icon.ico"