+++ CHANGES.rst
@@ -2,13 +2,21 @@
 Changelog
 =========
 
+Version 2.0
+===========
+*Released 2013-01-06*
+
+* Upgrade ``python-telegram-bot`` to 20.0
+* Drop ``ptbstats``
+* Move repository to GitHub
+
 Version 1.1
 ===========
 *Released 2020-08-17*
 
 **Enhancements:**
 
-* Include the `ptbstats <https://hirschheissich.gitlab.io/ptbstats/>`_ plugin for statistics
+* Include the `ptbstats <https://Bibo-Joshi.github.io/ptbstats/>`_ plugin for statistics
 
 Version 1.0.1
 =============